Fictional Pub&Beer webpage for reservations
This is a final lesson of a module from the HTML & CSS course that I'm watching.
It's a fictional webpage simulating a reservation form for a Pub&Beer restaurant.

The logo and the bokeh lights were provided by the teacher. I took my creative freedom to put other elements like the beer picture on the background and some final styling touches on the elements like the "reserve" button animation, the border styles, etc.

The original lesson proposal was only 3 form fields, and the  "reservation" field was meant to be Text type. But instead of that, I decided to transform the "reservation" field into 2 other fields "date" and "time". So I put a calendar and a time selector in HTML, with visual validation (CSS) according to the time interval rule (7pm - 23pm).